Zmieniono metodę SzescianPlusKwadrat(double a). Jest w stanie robić sześcian ze zmiennej typu double

+++ Github_Kalkulator_Z_Galeziami/Kalkulator.cs
@@ -21,7 +21,7 @@ namespace Github_Kalkulator_Z_Galeziami
 
         public static double SzescianPlusKwadrat(double a)
         {
-            return Szescian((int)a) + Kwadrat((int)a);
+            return Szescian(a) + Kwadrat((int)a);
         }
     }
 }
